Can now add lectures to the db
